Lift some API test fixtures up to run on E2E too
