+++ mathesar/tests/conftest.py
@@ -1,9 +1,16 @@
+from django.core.files import File
 """
 This inherits the fixtures in the root conftest.py
 """
 import pytest
 
+from sqlalchemy import text
+
+from db.schemas.operations.create import create_schema as create_sa_schema
+from db.schemas.utils import get_schema_oid_from_name, get_schema_name_from_oid
+from mathesar.imports.csv import create_table_from_csv
 from mathesar.models import Database
+from mathesar.models import Schema, DataFile
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -87,3 +94,39 @@ def data_types_csv_filename():
 @pytest.fixture(scope='session')
 def non_unicode_csv_filename():
     return 'mathesar/tests/data/non_unicode_files/utf_16_le.csv'
+
+
+@pytest.fixture
+def create_schema(engine, test_db_model):
+    """
+    Creates a schema factory, making sure to track and clean up new instances
+    """
+    function_schemas = {}
+
+    def _create_schema(schema_name):
+        if schema_name in function_schemas:
+            schema_oid = function_schemas[schema_name]
+        else:
+            create_sa_schema(schema_name, engine)
+            schema_oid = get_schema_oid_from_name(schema_name, engine)
+            function_schemas[schema_name] = schema_oid
+        schema_model, _ = Schema.current_objects.get_or_create(oid=schema_oid, database=test_db_model)
+        return schema_model
+    yield _create_schema
+
+    for oid in function_schemas.values():
+        # Handle schemas being renamed during test
+        schema = get_schema_name_from_oid(oid, engine)
+        with engine.begin() as conn:
+            conn.execute(text(f'DROP SCHEMA IF EXISTS "{schema}" CASCADE;'))
+
+
+@pytest.fixture
+def create_table(csv_filename, create_schema):
+    with open(csv_filename, 'rb') as csv_file:
+        data_file = DataFile.objects.create(file=File(csv_file))
+
+    def _create_table(table_name, schema='Patents'):
+        schema_model = create_schema(schema)
+        return create_table_from_csv(data_file, table_name, schema_model)
+    return _create_table
